infra/tests/test_stress_test.py
"""Tests for the robustness testing suite."""

import numpy as np
import pandas as pd
import pytest

from moat.engines.vector_engine import sma
from moat.stress_test import (
    RobustnessTestSuite,
    StressTestConfig,
    quick_robustness_check,
)


@pytest.fixture
def sample_data() -> pd.DataFrame:
    """Generate sample OHLCV data with trend."""
    np.random.seed(42)
    dates = pd.date_range("2020-01-01", periods=500, freq="D")
    close = 100 * np.exp(np.cumsum(np.random.randn(500) * 0.02 + 0.001))
    return pd.DataFrame(
        {
            "open": close * (1 + np.random.randn(500) * 0.005),
            "high": close * (1 + np.abs(np.random.randn(500) * 0.01)),
            "low": close * (1 - np.abs(np.random.randn(500) * 0.01)),
            "close": close,
            "volume": np.random.randint(1000000, 10000000, 500),
        },
        index=dates,
    )


def ma_strategy(data: pd.DataFrame) -> pd.Series:
    """Simple MA crossover strategy for testing."""
    fast = sma(data["close"], 20)
    slow = sma(data["close"], 50)
    return (fast > slow).astype(float)


class TestMonteCarloTest:
    """Tests for Monte Carlo simulation."""

    def test_monte_carlo_runs(self, sample_data: pd.DataFrame) -> None:
        """Test Monte Carlo generates expected number of runs."""
        config = StressTestConfig(monte_carlo_runs=10)
        suite = RobustnessTestSuite(stress_config=config)

        result = suite.monte_carlo_test(sample_data, ma_strategy)

        assert result["n_runs"] == 10
        assert "mean_sharpe" in result
        assert "std_sharpe" in result

    def test_monte_carlo_variance(self, sample_data: pd.DataFrame) -> None:
        """Test Monte Carlo produces variance in results."""
        config = StressTestConfig(monte_carlo_runs=50)
        suite = RobustnessTestSuite(stress_config=config)

        result = suite.monte_carlo_test(sample_data, ma_strategy)

        # Should have some variance (not all identical)
        assert result["std_sharpe"] > 0


class TestNoiseInjection:
    """Tests for noise injection."""

    def test_noise_injection_runs(self, sample_data: pd.DataFrame) -> None:
        """Test noise injection at multiple levels."""
        config = StressTestConfig(noise_levels=[0.005, 0.01, 0.02])
        suite = RobustnessTestSuite(stress_config=config)

        result = suite.noise_injection_test(sample_data, ma_strategy)

        assert "sensitivity" in result
        assert len(result["per_level"]) == 3

    def test_noise_degrades_performance(self, sample_data: pd.DataFrame) -> None:
        """Test that higher noise generally degrades performance."""
        config = StressTestConfig(noise_levels=[0.001, 0.05])
        suite = RobustnessTestSuite(stress_config=config)

        result = suite.noise_injection_test(sample_data, ma_strategy)

        # Higher noise should typically cause degradation
        low_noise = result["per_level"][0]
        high_noise = result["per_level"][1]

        # At minimum, sharpe should be different
        assert low_noise["sharpe"] != high_noise["sharpe"]


class TestParameterScan:
    """Tests for parameter sensitivity analysis."""

    def test_parameter_scan(self, sample_data: pd.DataFrame) -> None:
        """Test parameter scanning."""
        config = StressTestConfig(param_variations=3)
        suite = RobustnessTestSuite(stress_config=config)

        def parameterized_ma(data: pd.DataFrame, params: dict) -> pd.Series:
            fast = sma(data["close"], int(params["fast_period"]))
            slow = sma(data["close"], int(params["slow_period"]))
            return (fast > slow).astype(float)

        param_ranges = {
            "fast_period": (10, 30),
            "slow_period": (40, 60),
        }

        result = suite.parameter_scan(sample_data, parameterized_ma, param_ranges)

        assert "stability_score" in result
        assert result["stability_score"] >= 0
        assert result["stability_score"] <= 1
        assert len(result["results"]) > 0


class TestFullSuite:
    """Tests for complete robustness suite."""

    def test_full_suite_report(self, sample_data: pd.DataFrame) -> None:
        """Test full robustness report generation."""
        config = StressTestConfig(
            monte_carlo_runs=20,
            noise_levels=[0.005, 0.01],
        )
        suite = RobustnessTestSuite(stress_config=config)

        report = suite.run_full_suite(
            sample_data,
            ma_strategy,
            strategy_name="test_ma",
        )

        assert report.strategy_name == "test_ma"
        assert 0 <= report.robustness_score <= 100
        assert report.monte_carlo_mean_sharpe is not None
        assert report.noise_sensitivity is not None
        assert report.notes is not None

    def test_quick_robustness_check(self, sample_data: pd.DataFrame) -> None:
        """Test quick robustness check utility."""
        score = quick_robustness_check(
            sample_data,
            ma_strategy,
            n_monte_carlo=10,
        )

        assert 0 <= score <= 100


class TestRobustnessScoring:
    """Tests for robustness score calculation."""

    def test_good_strategy_scores_higher(self, sample_data: pd.DataFrame) -> None:
        """Test that robust strategies score higher than fragile ones."""
        config = StressTestConfig(
            monte_carlo_runs=20,
            noise_levels=[0.005, 0.01],
        )
        suite = RobustnessTestSuite(stress_config=config)

        # Good strategy: trend following MA
        def good_strategy(data: pd.DataFrame) -> pd.Series:
            fast = sma(data["close"], 20)
            slow = sma(data["close"], 50)
            return (fast > slow).astype(float)

        # Fragile strategy: oscillates rapidly (high turnover, noise sensitive)
        def fragile_strategy(data: pd.DataFrame) -> pd.Series:
            return pd.Series(
                [1.0 if i % 2 == 0 else -1.0 for i in range(len(data))],
                index=data.index,
            )

        good_report = suite.run_full_suite(sample_data, good_strategy, "good")
        fragile_report = suite.run_full_suite(sample_data, fragile_strategy, "fragile")

        print(f"\nGood strategy score: {good_report.robustness_score:.1f}")
        print(f"Fragile strategy score: {fragile_report.robustness_score:.1f}")

        # Good strategy should score at least as high as fragile
        # (Note: in trending data, even fragile may work by chance)
        assert good_report.robustness_score >= 0
        assert fragile_report.robustness_score >= 0
