infra/pyproject.toml
[project]
name = "moat"
version = "0.1.0"
description = "Quantitative Research & Execution Infrastructure"
readme = "README.md"
authors = [
    { name = "rmi11er", email = "reagan@outlook.ie" }
]
requires-python = ">=3.10"
dependencies = [
    # Data
    "pandas>=2.0.0",
    "polars>=0.20.0",
    "pyarrow>=14.0.0",
    # Config & Validation
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    # Data Sources
    "yfinance>=0.2.0",
    # File Watching
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "hypothesis>=6.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pandas-stubs>=2.0.0",
]

[project.scripts]
moat = "moat:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
